fix: 修复阅读器模型层编译错误
- TextColumn 添加 Hashable 协议支持
- TextPage 修复 Float 到 CGFloat 类型转换
- TextChapter 修复静态属性 empty 初始化问题

+++ Core/Reader/Models/TextPage.swift
@@ -101,13 +101,12 @@ class TextPage {
     }
     
     func draw(in view: ContentTextView, context: CGContext, relativeOffset: Float) {
-        // TODO: 实现页面绘制逻辑
         context.saveGState()
-        context.translateBy(x: 0, y: relativeOffset)
+        context.translateBy(x: 0, y: CGFloat(relativeOffset))
         
         for line in lines {
             context.saveGState()
-            context.translateBy(x: 0, y: line.lineTop)
+            context.translateBy(x: 0, y: CGFloat(line.lineTop))
             line.draw(in: view, context: context)
             context.restoreGState()
         }
